+++ crates/biome_cli/src/execute/migrate/eslint_any_rule_to_biome.rs
@@ -185,6 +185,14 @@ pub(crate) fn migrate_eslint_any_rule(
                 .get_or_insert(Default::default());
             rule.set_level(rule_severity.into());
         }
+        "@typescript-eslint/no-restricted-types" => {
+            if !options.include_nursery {
+                return false;
+            }
+            let group = rules.nursery.get_or_insert_with(Default::default);
+            let rule = group.no_restricted_types.get_or_insert(Default::default());
+            rule.set_level(rule_severity.into());
+        }
         "@typescript-eslint/no-this-alias" => {
             if !options.include_inspired {
                 results.has_inspired_rules = true;